Move notifications to its own folder in service package
